Made sunburstSeries more component-like
